pyproject.toml
[project]
name = "mcp-jira-python"
version = "0.1.0"
description = "MCP server for JIRA in Python"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [ "mcp>=1.2.1", "jira", "python-dotenv" ]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.8",
    "mypy>=1.13",
    "bandit>=1.7",
    "pre-commit>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_jira_python"]

[project.scripts]
jira-api = "mcp_jira_python:main"

# ============================================================================
# Ruff - Linting and Formatting
# ============================================================================
[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "PL",     # pylint
    "RUF",    # ruff-specific
]
ignore = [
    "PLR0913",  # too many arguments
    "PLR2004",  # magic value comparison
]

[tool.ruff.lint.isort]
known-first-party = ["mcp_jira_python"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ============================================================================
# Mypy - Type Checking
# ============================================================================
[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false  # MCP decorators aren't fully typed
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
files = ["src/mcp_jira_python"]

[[tool.mypy.overrides]]
module = ["jira.*", "mcp.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["mcp_jira_python.tools.*"]
# Tools have runtime checks for self.jira being set
disable_error_code = ["union-attr"]

# ============================================================================
# Bandit - Security Scanning
# ============================================================================
[tool.bandit]
exclude_dirs = ["tests", ".venv", "venv"]
skips = []

# ============================================================================
# Pytest - Testing
# ============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "unit: Unit tests (mocked, fast)",
    "integration: Integration tests (requires Jira connection)",
    "e2e: End-to-end tests (full MCP flow)",
    "slow: Slow running tests",
]
addopts = [
    "-v",
    "--strict-markers",
]

# Run only unit tests by default (fast, no external deps)
# Use: pytest -m unit           # unit tests only
# Use: pytest -m integration    # integration tests only
# Use: pytest -m e2e            # e2e tests only
# Use: pytest                   # all tests
# Use: pytest --cov             # all tests with coverage

# ============================================================================
# Coverage - Code Coverage
# ============================================================================
[tool.coverage.run]
source = ["src/mcp_jira_python"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
fail_under = 90
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"
